fixed typo

+++ standard_deviation.py
@@ -38,7 +38,7 @@ class Vector(object):
         self.total_elements = total_elements
         # self.a = numpy.random.randn(self.total_elements).astype(numpy.float32)
         # self.b = numpy.random.randn(self.total_elements).astype(numpy.float32)
-        self.a = numpy.array([i for i in range(10)])
+        self.a = numpy.array([i for i in range(10)]).astype(numpy.float32)
         self.result = numpy.zeros_like(self.a)
 
     def sum(self, blocks=None, threads=1):
